feat: Change trace_id to u128 to follow spec https://www.w3.org/TR/trace-context/#trace-id

+++ src/layer.rs
@@ -18,7 +18,7 @@ use crate::{
 
 #[derive(Copy, Clone)]
 struct TraceContext {
-    pub trace_id: u64,
+    pub trace_id: u128,
 }
 
 struct SpanContext {
